requirements.txt
Django>=4.2.0,<5.0.0
django-allauth>=0.54.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=9.5.0
python-dotenv>=1.0.0
whitenoise>=6.4.0
psycopg2-binary==2.9.9
psycopg==3.1.12
django-widget-tweaks>=1.4.12
django-countries>=7.5.1
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.11
django-storages>=1.13.2
django-taggit>=4.0.0
django-filter>=23.2
django-environ>=0.10.0
python-magic>=0.4.27
django-tinymce>=3.6.1
channels>=4.0.0
channels-redis>=4.1.0
djangorestframework>=3.14.0
dj-database-url>=2.1.0
social-auth-core>=4.5.0
social-auth-app-django>=5.4.0
six==1.16.0
nltk==3.8.1
scikit-learn==1.3.2